helper/exec: allow skipping cleanup for debugging
It's difficult to reproduce the context in which Otto executes Terraform
and Packer, because it generates a bunch of files that it uses to pass
into the tools, and then immediately removes them.

This adds support for `OTTO_SKIP_CLEANUP`, which, when used in concert
with `OTTO_LOG`, allows us to see precisely the commands that Otto runs
and reproduce them afterwards.

Example use case I wanted this for: get the context for `otto deploy`
and rerun the command flipping `apply` for `destroy` to work around the
lack of that feature at the moment.

Also tweaks the execHelper debug output to include CWD and avoid
double-printing the executable name.

+++ helper/packer/packer.go
@@ -35,7 +35,9 @@ func (p *Packer) Execute(commandRaw ...string) error {
 	if err != nil {
 		return err
 	}
-	defer os.Remove(varfile)
+	if execHelper.ShouldCleanup() {
+		defer os.Remove(varfile)
+	}
 
 	// The command must always be machine-readable. We use this
 	// exclusively to mirror the UI output.
